refactor: move the terminal WebSocket endpoints out of index.ts (#548 step 3e)
/ws, /ws/run, /ws/launch and /ws/codex — the composition root that ties the
spawners, the connection plumbing and the session decisions together — become
routes/ws-routes.ts. This is the last of the terminal machinery to leave; it only
became a contiguous block once the translation worker moved out.

index.ts keeps what it owns and passes it in: the http server these hang their
upgrade handler off, the origin check, the working/waiting flags, and the
spawners it built. Each endpoint's handler is a top-level function taking those
deps, so mounting stays a thin wiring step rather than one long closure.

Verified against main by booting both and comparing: five API endpoints, a real
command PTY (output and exit code), the /ws frame sequence, the 11 CLI arguments
claude is spawned with, and the launcher/run not-found error paths.
